+++ src/ltm.h
@@ -33,6 +33,7 @@ typedef enum {
   TM_LE,
   TM_CONCAT,
   TM_CALL,
+  TM_ENVIRONMENT,
   TM_N		/* number of elements in the enum */
 } TMS;